requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
structlog==24.1.0
kafka-python==2.0.2
pandas==2.2.2
numpy==1.26.4
httpx==0.27.0
redis==5.0.0